+++ custom-komapper-jdbc-change/src/main/kotlin/momosetkn/liquibase/kotlin/change/KomapperJdbcContext.kt
@@ -1,5 +1,7 @@
 package momosetkn.liquibase.kotlin.change
 
+import momosetkn.sql.DatasourceProxy
+import momosetkn.sql.NotCloseConnectionProxy
 import kotlin.reflect.full.declaredMemberProperties
 import kotlin.reflect.jvm.isAccessible
 
@@ -17,7 +19,7 @@ internal fun <R : Any> withKomapperJdbcContext(
 private fun liquibase.database.Database.toJavaxSqlDataSource(): DatasourceProxy {
     val liquibaseJdbcConnection = this.connection as liquibase.database.jvm.JdbcConnection
     val con = connectionProperty.get(liquibaseJdbcConnection) as java.sql.Connection
-    return DatasourceProxy(con)
+    return DatasourceProxy(NotCloseConnectionProxy(con))
 }
 
 private val connectionProperty = liquibase.database.jvm.JdbcConnection::class.declaredMemberProperties
